+++ app/Http/Controllers/RecordController.php
@@ -26,6 +26,11 @@ use Yajra\Datatables\Datatables;
 class RecordController extends Controller
 {
 
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     /**
      * Display a listing of the resource.
      *