+++ app/Http/Controllers/ZoneController.php
@@ -24,6 +24,10 @@ use Yajra\Datatables\Datatables;
 
 class ZoneController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
 
     /**
      * Display a listing of the resource.