+++ app/Http/Requests/UserUpdateRequest.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class UserUpdateRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        $user = $this->route('user');
+
+        return [
+
+            'name'     => 'required|string',
+            'email'    => 'required|email|unique:users,email,' . $user->id,
+            'password' => 'sometimes|alpha_num|min:6|confirmed',
+
+        ];
+    }
+}