+++ database/seeds/DatabaseSeeder.php
@@ -28,6 +28,7 @@ class DatabaseSeeder extends Seeder
     public function run()
     {
         $this->call(SettingsTableSeeder::class);
+        $this->call(UsersTableSeeder::class);
         $this->call(ServersTableSeeder::class);
         $this->call(ZonesTableSeeder::class);
     }