+++ resources/lang/en/general.php
@@ -33,7 +33,6 @@ return [
     'save'     => 'Save data',
     'inactive' => 'disabled',
     'help'     => 'Show / Hide help',
-    'logout'   => 'Sign out',
     'info'     => 'Info!',
     'warning'  => 'Warning!'
 ];