+++ resources/lang/en/user/messages.php
@@ -1,0 +1,43 @@
+<?php
+/**
+ * ProBIND v3 - Professional DNS management made easy.
+ *
+ * Copyright (c) 2016 by Paco Orozco <paco@pacoorozco.info>
+ *
+ * This file is part of some open source application.
+ *
+ * Licensed under GNU General Public License 3.0.
+ * Some rights reserved. See LICENSE, AUTHORS.
+ *
+ * @author      Paco Orozco <paco@pacoorozco.info>
+ * @copyright   2016 Paco Orozco
+ * @license     GPL-3.0 <http://spdx.org/licenses/GPL-3.0>
+ * @link        https://github.com/pacoorozco/probind
+ */
+
+return [
+    'already_exists' => 'User already exists!',
+    'does_not_exist' => 'User does not exist.',
+
+    'create' => [
+        'error'   => 'User was not created, please try again.',
+        'success' => 'User created successfully.',
+    ],
+
+    'update' => [
+        'error'   => 'User was not updated, please try again',
+        'success' => 'User updated successfully.',
+    ],
+
+    'delete' => [
+        'error'   => 'There was an issue deleting the user. Please try again.',
+        'success' => 'The user was deleted successfully.',
+        'invalid' => 'You can\'t delete yourself.',
+    ],
+
+    'activity' => [
+        'created' => 'New User <strong>:username</strong> was created with admin role.',
+        'updated' => 'User <strong>:username</strong> was modified.',
+        'deleted' => 'User <strong>:username</strong> was deleted.'
+    ]
+];