+++ tests/ZoneHttpTest.php
@@ -23,6 +23,14 @@ class ZoneHttpTest extends TestCase
 
     use DatabaseMigrations;
 
+    public function setUp()
+    {
+        parent::setUp();
+
+        $user = factory(\App\User::class)->create();
+        $this->actingAs($user);
+    }
+
     /**
      * Test a successful new Master Zone creation
      */